# perf: change ripmd160 from class to struct to avoid allocations.