--- a/AerospikeClientProxy/AerospikeClientProxy.csproj
+++ b/AerospikeClientProxy/AerospikeClientProxy.csproj
@@ -25,6 +25,8 @@
     <OutputType>Library</OutputType>
     <StartupObject />
     <RootNamespace>Aerospike.Client</RootNamespace>
+    <Configurations>Debug;Release;Debug_Unix;Release_Unix</Configurations>
+    <Platforms>AnyCPU</Platforms>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -33,6 +35,17 @@
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
     <NoWarn>1701;1702;CS1591</NoWarn>
+  </PropertyGroup>
+
+  <PropertyGroup Condition=" '$(Configuration)' == 'Debug_Unix' ">
+    <NoWarn>1701;1702;CS1591</NoWarn>
+    <DebugSymbols Condition=" '$(DebugSymbols)' == '' ">true</DebugSymbols>
+    <Optimize Condition=" '$(Optimize)' == '' ">false</Optimize>
+  </PropertyGroup>
+
+  <PropertyGroup Condition=" '$(Configuration)' == 'Release_Unix' ">
+    <NoWarn>1701;1702;CS1591</NoWarn>
+    <Optimize Condition=" '$(Optimize)' == '' ">true</Optimize>
   </PropertyGroup>
 
   <PropertyGroup>