--- a/AerospikeTest/AerospikeTest.csproj
+++ b/AerospikeTest/AerospikeTest.csproj
@@ -23,7 +23,9 @@
 
     <Authors>Aerospike</Authors>
 
-    <Configurations>Debug;Release;Coverage</Configurations>
+    <Configurations>Debug;Release;Coverage;Debug_Unix;Release_Unix</Configurations>
+
+    <Platforms>AnyCPU</Platforms>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -36,6 +38,17 @@
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
     <DefineConstants>TRACE;NETCORE</DefineConstants>
+  </PropertyGroup>
+
+  <PropertyGroup Condition=" '$(Configuration)' == 'Debug_Unix' ">
+    <DefineConstants>TRACE;DEBUG;NETCORE</DefineConstants>
+    <DebugSymbols Condition=" '$(DebugSymbols)' == '' ">true</DebugSymbols>
+    <Optimize Condition=" '$(Optimize)' == '' ">false</Optimize>
+  </PropertyGroup>
+
+  <PropertyGroup Condition=" '$(Configuration)' == 'Release_Unix' ">
+    <DefineConstants>TRACE;NETCORE</DefineConstants>
+    <Optimize Condition=" '$(Optimize)' == '' ">true</Optimize>
   </PropertyGroup>
   <PropertyGroup>
     <StartupObject />